requirements.txt
pandas
numpy
folium
matplotlib
tensorflow
gensim
nltk
pysastrawi
scikit-learn
fastapi
uvicorn
pydantic
thompson_sampling
